feat: up dependencies
