+++ example/app/wagmiConfig.ts
@@ -1,7 +1,10 @@
 import { connectorsForWallets } from "@rainbow-me/rainbowkit";
 import { createConfig } from "wagmi";
 import { hardhat, optimismSepolia } from "viem/chains";
-import { metaMaskWallet } from "@rainbow-me/rainbowkit/wallets";
+import {
+  metaMaskWallet,
+  walletConnectWallet,
+} from "@rainbow-me/rainbowkit/wallets";
 import { createClient, http } from "viem";
 import { rainbowkitBurnerWallet } from "burner-connector";
 
@@ -13,7 +16,7 @@ import { rainbowkitBurnerWallet } from "burner-connector";
   [optimismSepolia.id]: `https://opt-sepolia.g.alchemy.com/v2/${alchemyAPIKey}`,
 }; */
 
-const wallets = [metaMaskWallet, rainbowkitBurnerWallet];
+const wallets = [metaMaskWallet, walletConnectWallet, rainbowkitBurnerWallet];
 const walletConnectProjectID = "3a8170812b534d0ff9d794f19a901d64";
 const wagmiConnectors = connectorsForWallets(
   [
@@ -34,7 +37,6 @@ export const chains = [optimismSepolia, hardhat] as const;
 export const wagmiConfig = createConfig({
   chains: chains,
   connectors: wagmiConnectors,
-  ssr: true,
   client({ chain }) {
     return createClient({ chain, transport: http() });
   },
