+++ example/package.json
@@ -9,14 +9,14 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@rainbow-me/rainbowkit": "2.2.9",
-    "@tanstack/react-query": "^5.28.6",
+    "@rainbow-me/rainbowkit": "2.2.10",
+    "@tanstack/react-query": "^5.100.9",
     "burner-connector": "workspace:*",
     "next": "14.2.3",
     "react": "^18",
     "react-dom": "^18",
     "react-hot-toast": "^2.4.1",
-    "viem": "2.39.0",
+    "viem": "2.48.4",
     "wagmi": "2.19.5"
   },
   "devDependencies": {
