+++ packages/burner-connector/package.json
@@ -36,17 +36,17 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
-    "@eslint/js": "^9.1.1",
-    "eslint": "^8",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "prettier": "3.2.5",
+    "@eslint/js": "^9.39.0",
+    "eslint": "^9.39.0",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-prettier": "^5.5.5",
+    "prettier": "3.8.3",
     "typescript": "^5.4.5",
-    "typescript-eslint": "^7.7.1"
+    "typescript-eslint": "^8.59.1"
   },
   "dependencies": {
-    "@rainbow-me/rainbowkit": "2.2.9",
-    "viem": "2.39.0",
+    "@rainbow-me/rainbowkit": "2.2.10",
+    "viem": "2.48.4",
     "wagmi": "2.19.5"
   }
 }
